move bot to src/bot
