+++ src/bot/__tests__/__fixtures__/messagesContentSupplier/correctData.json
