+++ src/bot/__tests__/contentSupplier.test.js
@@ -1,5 +1,5 @@
 const moment = require('moment');
-const { onMessage } = require('../src/actions');
+const { onMessage } = require('../actions');
 const { sendTestMessage, createMockedSlimbot } = require('./lib/sendMessageHelper');
 
 const { describeDBSetupTeardown } = require('./lib/dbHelper');
