+++ src/bot/__tests__/lib/sendMessageHelper.js
