+++ src/bot/statistics/contentSupplierStat.js
@@ -1,6 +1,6 @@
 const moment = require('moment');
 const proschet = require('proschet').default;
-const { dbClient } = require('../dbClient');
+const { dbClient } = require('../../dbClient');
 const { getFullUserName } = require('../utils/render');
 
 moment.locale('ru');
