+++ src/bot/statistics/index.js
