+++ src/bot/utils/render.js
